refactor(engine): extract bpmn transformer
